+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+- Tables (fmt.table)
 
 ## [0.2.2] - TBD
 - Separate methods from index